+++ packages/@ionic/cli-utils/src/lib/hooks.ts
@@ -13,7 +13,6 @@ export class Hook<T, U> implements IHook<T, U> {
 }
 
 export class HookEngine implements IHookEngine {
-
   private hooks = new Map<string, IHook<any, any>[]>();
 
   register<T, U>(source: string, hook: string, listener: (args?: T) => Promise<U>) {
@@ -26,20 +25,6 @@ export class HookEngine implements IHookEngine {
     return Promise.all(registeredHooks.map(h => h.fire(args)));
   }
 
-  getSources(hook: string) {
-    return [...new Set(this.getRegistered(hook).map(h => h.source))];
-  }
-
-  hasSources(hook: string, sources: string[]) {
-    return sources.filter(s => this.getSources(hook).includes(s)).length > 0;
-  }
-
-  deleteSource(source: string) {
-    for (const [ hookName, hooks ] of this.hooks.entries()) {
-      this.hooks.set(hookName, hooks.filter(h => h.source !== source));
-    }
-  }
-
   getRegistered<T, U>(hook: string): IHook<T, U>[] {
     let registeredHooks = this.hooks.get(hook);
     if (!registeredHooks) {
@@ -49,5 +34,4 @@ export class HookEngine implements IHookEngine {
 
     return registeredHooks;
   }
-
 }