+++ packages/@ionic/cli-utils/src/lib/integrations/cordova/utils.ts
@@ -25,6 +25,10 @@ export function filterArgumentsForCordova(metadata: CommandMetadata, options: Co
   m.options.push(...globalCordovaOpts);
 
   const results = filterCommandLineOptionsByGroup(m, options, OptionGroup.Cordova);
+
+  results['target'] = results['cordova-target'];
+  delete results['cordova-target'];
+
   const args = unparseArgs(results, { useEquals: false, allowCamelCase: true });
   const i = args.indexOf('--');